# Optimize summarygroup endpoint