--- a/treeherder/webapp/api/groups.py
+++ b/treeherder/webapp/api/groups.py
@@ -8,7 +8,7 @@ from rest_framework import generics
 from rest_framework.response import Response
 
 from treeherder.model.models import JobLog
-from treeherder.webapp.api.serializers import GroupNameSerializer
+from collections import defaultdict
 
 logger = logging.getLogger(__name__)
 
@@ -18,7 +18,6 @@ class SummaryByGroupName(generics.ListAPIView):
     This yields group names/status summary for the given group and day.
     """
 
-    serializer_class = GroupNameSerializer
     queryset = None
 
     def list(self, request):
@@ -51,8 +50,10 @@ class SummaryByGroupName(generics.ListAPIView):
             .filter(
                 job__repository_id__in=(1, 77),
                 job__job_type__name__startswith='test-',
-                groups__name__isnull=False,
                 group_result__status__in=(1, 2),
+            )
+            .exclude(
+                groups__name='',
             )
             .annotate(
                 job_count=Count('job_id'),
@@ -61,46 +62,48 @@ class SummaryByGroupName(generics.ListAPIView):
                     When(group_result__status=2, then=Value("testfailed")),
                 ),
             )
-            .values(
-                'job_count',
-                'job__job_type__name',
-                'job__failure_classification_id',
+            .values_list(
                 'groups__name',
+                'job__job_type__name',
                 'result',
+                'job__failure_classification_id',
+                'job_count',
             )
             .order_by('groups__name')
         )
 
-        serializer = self.get_serializer(self.queryset, many=True)
+        # Reference job types in a separated list
+        job_type_names = set()
+        # Group items by group name, type name, result and classification
+        summary = defaultdict(lambda: defaultdict(lambda: defaultdict(lambda: defaultdict(int))))
+        for item in self.queryset.all():
+            group_name, type_name, result, classification, job_count = item
 
-        summary = {}
-        job_type_names = []
-        for item in serializer.data:
-            # TODO: consider stripping out some types; mostly care about FBC vs Intermittent
-            classification = item['failure_classification']
-            result = item["result"]
+            # Strip a possible number suffix
+            name, suffix = type_name.rsplit('-', maxsplit=1)
+            if suffix.isdigit():
+                type_name = name
 
-            if item['job_type_name'] not in job_type_names:
-                job_type_names.append(item['job_type_name'])
-            if item['group_name'] not in summary:
-                summary[item['group_name']] = {}
-            if item['job_type_name'] not in summary[item['group_name']]:
-                summary[item['group_name']][item['job_type_name']] = {}
-            if result not in summary[item['group_name']][item['job_type_name']]:
-                summary[item['group_name']][item['job_type_name']][result] = {}
-            if classification not in summary[item['group_name']][item['job_type_name']][result]:
-                summary[item['group_name']][item['job_type_name']][result][classification] = 0
-            summary[item['group_name']][item['job_type_name']][result][classification] += item[
-                'job_count'
-            ]
+            job_type_names.add(type_name)
+            summary[group_name][type_name][result][classification] += job_count
 
-        data = {'job_type_names': job_type_names, 'manifests': []}
-        for m in summary.keys():
+        # Cast job types as a list, to use their index as reference in manifests
+        job_type_names = sorted(job_type_names)
+
+        manifests = []
+        for group, types in summary.items():
             mdata = []
-            for d in summary[m]:
-                for r in summary[m][d]:
-                    for c in summary[m][d][r]:
-                        mdata.append([job_type_names.index(d), r, int(c), summary[m][d][r][c]])
-            data['manifests'].append({m: mdata})
+            for t_name, results in types.items():
+                for result, classifications in results.items():
+                    for classif, job_count in classifications.items():
+                        mdata.append(
+                            [job_type_names.index(t_name), result, int(classif), job_count]
+                        )
+            manifests.append({group: mdata})
 
-        return Response(data=data)
+        return Response(
+            data={
+                'job_type_names': job_type_names,
+                'manifests': manifests,
+            }
+        )