--- a/treeherder/webapp/api/serializers.py
+++ b/treeherder/webapp/api/serializers.py
@@ -319,11 +319,12 @@ class JobTypeNameField(serializers.Field):
 
 
 class GroupNameSerializer(serializers.ModelSerializer):
-    group_name = serializers.CharField(source="job_log__groups__name")
-    job_type_name = JobTypeNameField(source="job_type__name")
-    group_status = serializers.CharField(source="job_log__group_result__status")
-    failure_classification = serializers.CharField(source="failure_classification_id")
+    group_name = serializers.CharField(source="groups__name")
+    job_type_name = JobTypeNameField(source="job__job_type__name")
+    group_status = serializers.CharField(source="group_result__status")
+    failure_classification = serializers.CharField(source="job__failure_classification_id")
     job_count = serializers.IntegerField()
+
 
     class Meta:
         model = models.JobLog