Work on `groupsummary` endpoint performance:
* Base the query on the JobLog table
* Apply filters/transformation within the query when possible
* Avoid using a serializer to handle query output
* Clean some Python code

I got interesting performance improvements with different data size:
* `master`
  * Small : 619 ms ± 12.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * Medium : 1.88 s ± 99.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * Large : 12.2 s ± 412 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
* `61e2076`
  * Small : 481 ms ± 38.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * Medium : 536 ms ± 14.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * Large : 4.32 s ± 396 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  
  The results were ordered differently, so I sorted the first index list of type names so the payload is a bit cleaner.
  The references in groups seems to match those types correctly.