@vrigal 2023-09-19
Costs (explain analyze) from pgadmin4 are pretty close (respectively 170k and 140k) but in practice the query is much faster to run.

* `master`
![master](https://github.com/mozilla/treeherder/assets/45713745/f744299c-0ecf-4fb0-ad0c-7d425b03a303)

* `61e2076`
![61e2076](https://github.com/mozilla/treeherder/assets/45713745/5801bad3-cbc5-4605-9cdc-f6529ad43f13)