@Archaeopteryx 2023-09-19
@jmaher Valentin made substantial improvements to the response time of the groupsummary queries. Can these land?